Renamed things. New threadpool compiles in nicely. Time to test it out. Also need to make it's compilation dummifiable.
