+++ c++/logging.hpp
