+++ c++/timer.cpp
@@ -5,8 +5,8 @@
 #include <unordered_map>
 #include <unistd.h>
 
-#include "logging.h"
-#include "timer.h"
+#include "logging.hpp"
+#include "timer.hpp"
 
 
 // Aonymous namespace
